t9_fastAPI/fast_api/pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["svishs <vishnichenko.sergey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
gunicorn = "^21.2.0"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
jinja2 = "^3.1.3"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
